boxing, enumeração, datas e valores monetários

+++ banco/src/com/mfarias25/banco/modelo/pagamento/DocumentoPagavel.java
@@ -1,8 +1,10 @@
 package com.mfarias25.banco.modelo.pagamento;
 
+import java.math.BigDecimal;
+
 public interface DocumentoPagavel {
 
-    double getValorTotal();
+    BigDecimal getValorTotal();
     boolean estaPago();
     void quitarPagamento();
     default void imprimirRecibo() {
